+++ README.md
@@ -1,2 +1,15 @@
 # Ostrichland
+
 Feed this Bad Boi. Ruby Text RPG.
+
+Install the dependencies:
+
+```bash
+$ bundle
+```
+
+Play the game:
+
+```bash
+$ ruby src/main.rb
+```
