--- src/Event/NPC/npc.rb
@@ -1,16 +1,0 @@
-require_relative '../event.rb'
-
-class NPC < Event
-
-  def initialize
-    super
-    @command = "talk"
-  end
-
-  def run(entity)
-    puts "Hello!"
-  end
-
-  attr_accessor :name
-
-end
