--- src/driver.rb
@@ -1,21 +1,0 @@
-require_relative "command.rb"
-
-def run_driver(player)
-  system("clear")
-
-  prompt(player)
-  input = player_input
-
-  while (input.casecmp("quit") != 0)
-    interpret_command(input, player)
-    input = player_input
-  end
-
-end
-
-def player_input
-  print "> "
-  input = gets.chomp
-  puts "\n"
-  return input
-end
