--- test/test_all.rb
@@ -1,7 +1,0 @@
-# Include all tests in this file.
-# Alphabetical order.
-
-require_relative "test_entity.rb"
-require_relative "test_item.rb"
-require_relative "test_map.rb"
-require_relative "test_plant.rb"
